+++ src/main/java/com/gajob/service/study/StudyServiceImpl.java
@@ -75,8 +75,17 @@ public class StudyServiceImpl implements StudyService {
   // 게시물 수정
   @Transactional
   public StudyReadDto update(Long postId, StudyDto studyDto) {
+    User user = userRepository.findOneWithAuthoritiesByEmail(
+        SecurityUtil.getCurrentUsername().get()).get();
+
     Study study = studyRepository.findById(postId)
         .orElseThrow(() -> new CustomException(ErrorCode.POST_ID_NOT_EXIST));
+
+    // 현재 로그인한 유저와 게시물 작성자의 이메일이 일치하지 않을 경우, 에러 발생
+    if (!(study.getUser().getEmail().equals(user.getEmail()))) {
+      throw new CustomException(ErrorCode.NO_ACCESS_RIGHTS);
+    }
+
     study.update(studyDto.getTitle(), studyDto.getContent(), studyDto.getStudyCategory(),
         studyDto.getArea(), studyDto.getMinPeople(), studyDto.getMaxPeople(),
         studyDto.getStartDate(),
@@ -90,8 +99,17 @@ public class StudyServiceImpl implements StudyService {
   // 게시물 삭제
   @Transactional
   public String delete(Long postId) {
+    User user = userRepository.findOneWithAuthoritiesByEmail(
+        SecurityUtil.getCurrentUsername().get()).get();
+
     Study study = studyRepository.findById(postId)
         .orElseThrow(() -> new CustomException(ErrorCode.POST_ID_NOT_EXIST));
+
+    // 현재 로그인한 유저와 게시물 작성자의 이메일이 일치하지 않을 경우, 에러 발생
+    if (!(study.getUser().getEmail().equals(user.getEmail()))) {
+      throw new CustomException(ErrorCode.NO_ACCESS_RIGHTS);
+    }
+
     studyRepository.delete(study);
 
     return "posts-delete";
